pyproject.toml
[tool.poetry]
name = "iqstocker"
version = "0.1.0"
description = "IQStocker бот и связанные сервисы"
authors = ["IQStocker Team <support@iqstocker.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
aiogram = "3.13.1"
python-telegram-bot = "21.7"
sqlalchemy = "2.0.36"
alembic = "1.13.3"
psycopg2-binary = "2.9.10"
asyncpg = "0.29.0"
aiosqlite = "0.20.0"
redis = "5.2.0"
pandas = "2.2.3"
numpy = "2.1.3"
chardet = "5.2.0"
scikit-learn = "1.5.2"
openai = "1.54.4"
anthropic = "0.34.2"
google-generativeai = "0.8.3"
cryptography = "42.0.5"
aiohttp = "3.10.11"
beautifulsoup4 = "4.12.3"
playwright = "1.49.1"
lxml = "5.3.0"
flask = "3.1.0"
flask-admin = "1.6.1"
flask-sqlalchemy = "3.1.1"
flask-wtf = "1.2.2"
sqladmin = "0.19.0"
plotly = "5.18.0"
ydata-profiling = "4.17.0"
python-multipart = "0.0.6"
jinja2 = "3.1.3"
itsdangerous = ">=2.2"
pillow = "10.0.0"
fastapi = "0.115.6"
uvicorn = "0.32.1"
celery = "5.4.0"
apscheduler = "3.10.4"
dramatiq = "1.18.0"
python-dotenv = "1.0.1"
pydantic = ">=2.4.1,<2.10"
pydantic-settings = "2.6.1"
structlog = "24.4.0"
sentry-sdk = { version = "2.18.0", extras = ["flask"] }
httpx = ">=0.27,<0.28"
supabase = ">=2.8.0"
psutil = "5.9.8"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-asyncio = "0.24.0"
black = "24.10.0"
isort = "5.13.2"
flake8 = "7.1.1"

[build-system]
requires = ["poetry-core>=1.8.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

